Extrai fechamento do pipeline de hipótese
